pyproject.toml
[tool.poetry]
name = "fd-zenmoney"
version = "0.1.0"
description = ""
authors = ["fdoooch <fdoooch@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.3"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
